Change tactic of inititialization.
My original approach of building docker images to do the configuration had a
flaw.  It did not account for the fact that ec2 instances apparently can't
resolve route53 dns names.  To work around this, I am readding the ports
for the server and agent containers.  This will allow me to communicate
with the consul instances in a shell script using the bound port.

This will requre me to place firewall rules to secure port 8500 from being
accessed from outside the swarm.